### 2019-04-19 #1394 @whedon
[PRE REVIEW]: SCONE: Open Source Software for Predictive Simulation of Biological Motion
**Submitting author:** @tgeijten (<a href="http://orcid.org/0000-0001-8352-8107">Thomas Geijtenbeek</a>)
**Repository:** <a href="https://github.com/opensim-org/SCONE" target ="_blank">https://github.com/opensim-org/SCONE</a>
**Version:** v1.0.0
