### 2019-04-19 #1394 @tgeijten
Thanks! The bullets points (*) don't show up correctly. Are they supported?
