### 2019-04-19 #1394 @arfon
The Markdown parser is a little fussy and needs there to be whitespace before the first bullet. This PR should fix the rendering: https://github.com/opensim-org/SCONE/pull/47
