### 2019-04-20 #1394 @tgeijten
Great, thanks! I merged the PR.
