### 2019-04-22 #1394 @Kevin-Mattheus-Moerman
@modenaxe @chrisdembia @moorepants @trallard @demotu @nicoguaro would you be interested in reviewing this work for JOSS?
