### 2019-04-22 #1394 @nicoguaro
@Kevin-Mattheus-Moerman, currently I am reviewing another work for JOSS. So, I will decline this one.
