### 2019-04-22 #1394 @demotu
yes, I can review it

@demotu <https://github.com/demotu>
-
Marcos Duarte
http://demotu.org/

On Mon, Apr 22, 2019 at 10:54 AM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @modenaxe <https://github.com/modenaxe> @chrisdembia
> <https://github.com/chrisdembia> @moorepants
> <https://github.com/moorepants> @trallard <https://github.com/trallard>
> @demotu <https://github.com/demotu> @nicoguaro
> <https://github.com/nicoguaro> would you be interested in reviewing this
> work for JOSS?
>
