### 2019-04-24 c4ae4f4
U: sconecmd: improved help messages, code cleanup
