### 2019-04-24 0a297f1
U: SconeStudio.cpp include file cleanup
