### 2019-04-25 #1394 @modenaxe
I can review it. Also, I have tried the alpha version of SCONE few months ago - just disclosing this in case it constitutes a problem -.
