+++ src/sconelib/scone/core/HasSignature.cpp
@@ -24,10 +24,10 @@ namespace scone
 		xo::replace_str( signature_postfix, "DATE_TIME_EXACT", GetDateTimeExactAsString() );
 		xo::replace_str( signature_prefix, "DATE_TIME", GetDateTimeAsString() );
 		xo::replace_str( signature_postfix, "DATE_TIME", GetDateTimeAsString() );
-		xo::replace_str( signature_prefix, "SCONE_BUILD", to_str( GetSconeVersion().build ) );
-		xo::replace_str( signature_postfix, "SCONE_BUILD", to_str( GetSconeVersion().build ) );
-		xo::replace_str( signature_prefix, "SCONE_VERSION", to_str( GetSconeVersion().build ) );
-		xo::replace_str( signature_postfix, "SCONE_VERSION", to_str( GetSconeVersion().build ) );
+		xo::replace_str( signature_prefix, "SCONE_BUILD", to_str( GetSconeVersion().build_ ) );
+		xo::replace_str( signature_postfix, "SCONE_BUILD", to_str( GetSconeVersion().build_ ) );
+		xo::replace_str( signature_prefix, "SCONE_VERSION", to_str( GetSconeVersion().build_ ) );
+		xo::replace_str( signature_postfix, "SCONE_VERSION", to_str( GetSconeVersion().build_ ) );
 	}
 
 	HasSignature::~HasSignature()
