### 2019-04-29 b389832
F*: Fixed documentation of JointLoadMeasure and ReactionForceMeasure F: inheritance from RangePenalty is now public in JointLoadMeasure / ReactionForceMeasure
