### 2019-04-29 84515f3
R: renamed xo::path::string() to xo::path::str() to prevent GCC error
