+++ src/sconestudio/main.cpp
@@ -45,7 +45,7 @@ int main( int argc, char *argv[] )
 		// init logging
 		xo::path log_file = scone::GetSettingsFolder() / "log" / xo::path( xo::get_date_time_str( "%Y%m%d_%H%M%S" ) + ".log" );
 		xo::log::file_sink file_sink( xo::log::debug_level, log_file );
-		SCONE_THROW_IF( !file_sink.good(), "Could not create file " + log_file.string() );
+		SCONE_THROW_IF( !file_sink.good(), "Could not create file " + log_file.str() );
 		xo::log::debug( "Created log file ", log_file );
 #ifdef _DEBUG
 		xo::log::stream_sink console_log_sink( xo::log::trace_level, std::cout );
