### 2019-04-29 bae8bf8
U*: Many documentation updates
