### 2019-04-30 a80a42e
N: Initial stub class with sol/lua interpreter
