### 2019-04-30 #1394 @kyleniemeyer
@tgeijten Generally JOSS likes to have at least two reviewers, though we can also have one if the editor serves as a second reviewer.
