+++ src/sconelib/scone/sconelib_config.h
@@ -18,17 +18,24 @@
 #include "sconeopensim4/ModelOpenSim4.h"
 #endif
 
+#ifdef SCONE_LUA
+#include "sconelua/ScriptController.h"
+#endif
+
 namespace scone
 {
 	void RegisterModels()
 	{
 #ifdef SCONE_OPENSIM_3
-	GetModelFactory().register_type< scone::ModelOpenSim3 >( "Simbody" );
-	GetModelFactory().register_type< scone::ModelOpenSim3 >( "OpenSim3Model" );
-	GetModelFactory().register_type< scone::ModelOpenSim3 >( "OpenSimModel" );
+	GetModelFactory().register_type< ModelOpenSim3 >( "Simbody" );
+	GetModelFactory().register_type< ModelOpenSim3 >( "OpenSim3Model" );
+	GetModelFactory().register_type< ModelOpenSim3 >( "OpenSimModel" );
 #endif
 #ifdef SCONE_OPENSIM_4
-	GetModelFactory().register_type< scone::ModelOpenSim4 >( "OpenSim4Model" );
+	GetModelFactory().register_type< ModelOpenSim4 >( "OpenSim4Model" );
+#endif
+#ifdef SCONE_LUA
+	GetControllerFactory().register_type< ScriptController >( "ScriptController" );
 #endif
 	}
 }
