+++ src/sconelib/sconelua/ScriptController.cpp
@@ -10,17 +10,23 @@
 #include "scone/model/Model.h"
 #include "scone/model/Actuator.h"
 #include "scone/core/string_tools.h"
+#include "LuaScript.h"
 
 namespace scone
 {
 	ScriptController::ScriptController( const PropNode& props, Params& par, Model& model, const Location& loc ) :
-	Controller( props, par, model, loc )
-	{
-		lua_.script( "x = 1" );
-	}
+	Controller( props, par, model, loc ),
+	script_( new LuaScript( props, par, model ) )
+	{}
+
+	ScriptController::~ScriptController()
+	{}
 
 	bool ScriptController::ComputeControls( Model& model, double timestamp )
 	{
+		if ( !script_->Run() )
+			return true;
+
 		return false;
 	}
 
