### 2019-05-01 055d378
U: Great progress, sol / lua work super smooth and easy! Ready to decide on full implementation.
