### 2019-05-02 5ff9eb9
Fixed typo in paper.md
