### 2019-05-02 ab8d55d
U: major updates, ScriptController and ScriptMeasure work great, created working high jump example.
