+++ src/sconelib/sconelua/ScriptController.cpp
@@ -11,27 +11,36 @@
 #include "scone/model/Actuator.h"
 #include "scone/core/string_tools.h"
 #include "LuaScript.h"
+#include "scone_lua_api.h"
 
 namespace scone
 {
 	ScriptController::ScriptController( const PropNode& props, Params& par, Model& model, const Location& loc ) :
 	Controller( props, par, model, loc ),
 	script_( new LuaScript( props, par, model ) )
-	{}
+	{
+		init_ = script_->GetFunction( "init" );
+		update_ = script_->GetFunction( "update" );
+
+		lua_params lp( par );
+		lua_model lm( model );
+		init_( &lm, &lp );
+
+		model.AddExternalResource( script_->GetFile() );
+	}
 
 	ScriptController::~ScriptController()
 	{}
 
 	bool ScriptController::ComputeControls( Model& model, double timestamp )
 	{
-		if ( !script_->Run() )
-			return true;
-
+		lua_model lm( model );
+		update_( &lm );
 		return false;
 	}
 
 	String ScriptController::GetClassSignature() const
 	{
-		return "S";
+		return "SC";
 	}
 }
