+++ src/sconelib/sconeopensim3/ModelOpenSim3.cpp
@@ -240,7 +240,7 @@ namespace scone
 
 		// create and initialize controllers
 		CreateControllers( props, par );
-		log::info( "Created model ", GetName(), "; dofs=", GetDofs().size(), " muscles=", GetMuscles().size(), " mass=", GetMass() );
+		log::debug( "Created model ", GetName(), "; dofs=", GetDofs().size(), " muscles=", GetMuscles().size(), " mass=", GetMass() );
 	}
 
 	ModelOpenSim3::~ModelOpenSim3() {}
