+++ src/sconelib/sconelua/ScriptController.h
@@ -16,5 +16,7 @@ namespace scone
 		virtual String GetClassSignature() const override;
 
 		u_ptr< class LuaScript > script_;
+		std::function<void( struct lua_model*, struct lua_params* )> init_;
+		std::function<void( struct lua_model* )> update_;
 	};
 }

[…]
