### 2019-05-06 39c3005
U: added *.lua to open file options
