### 2019-05-06 20e7873
U: added SCONE_ERROR for error message without function name
