+++ src/sconelib/scone/core/Exception.h
@@ -18,11 +18,11 @@
 namespace scone
 {
 	struct RuntimeException : public std::runtime_error {
-		RuntimeException(const std::string& msg) : runtime_error(msg) { }; // { std::cout << "scone::RuntimeException in " << msg << std::endl; }
+		RuntimeException( const std::string& msg ) : runtime_error( msg ) { };
 	};
 
 	struct LogicException : public std::logic_error {
-		LogicException(const std::string& msg) : logic_error(msg) { }; // { std::cout << "scone::LogicException in " << msg << std::endl; }
+		LogicException( const std::string& msg ) : logic_error( msg ) { };
 	};
 }
 
@@ -40,14 +40,21 @@ namespace scone
 
 #define SCONE_ASSERT( _expression_ ) \
 	if (!(_expression_)) throw scone::RuntimeException( "Assertion Failure in " + std::string( __func__ ) + "(): "#_expression_ );
-
 #define SCONE_ASSERT_MSG( _expression_, message ) \
 	if (!(_expression_)) throw scone::RuntimeException( "Assertion Failure in " + std::string( __func__ ) + "(): "#_expression_" (" + String( message ) + ")" );
 
-#define SCONE_THROW( message ) throw scone::RuntimeException(std::string( __func__ ) + "(): " + std::string( message ) )
+#define SCONE_THROW( message ) \
+	throw scone::RuntimeException( std::string( __func__ ) + "(): " + std::string( message ) )
 #define SCONE_THROW_IF( _condition_, _message_ ) \
-    do { \
-        if ( _condition_ ) throw scone::RuntimeException( std::string( __func__ ) + "(): " + std::string( _message_ ) ); } while (false);
+	if ( _condition_ ) throw scone::RuntimeException( std::string( __func__ ) + "(): " + std::string( _message_ ) );
 
 #define SCONE_CHECK_RANGE( _value_, _min_, _max_ ) \
 	if ( _value_ < _min_ || _value_ > _max_ ) throw scone::RuntimeException( #_value_" must be a value between " + to_str( _min_ ) + " and " + to_str( _max_ ) );
+
+#define SCONE_ERROR( _message_ ) \
+	throw scone::RuntimeException( std::string( _message_ ) )
+#define SCONE_ERROR_IF( _condition_, _message_ ) \
+	if ( _condition_ ) throw scone::RuntimeException( std::string( _message_ ) );
+
+#define SCONE_CATCH_AND_THROW( _message_ ) \
+	catch ( const std::exception& e ) { throw scone::RuntimeException( std::string( _message_ ) + e.what() ); }
