+++ src/sconelib/sconelua/ScriptMeasure.cpp
@@ -2,6 +2,7 @@
 
 #include "LuaScript.h"
 #include "scone_lua_api.h"
+#include "scone/core/Log.h"
 
 namespace scone
 {
@@ -15,7 +16,15 @@ namespace scone
 		result_ = script_->GetFunction( "result" );
 
 		lua_model lm( model );
-		init_( &lm );
+		try
+		{
+			lua_model lm( model );
+			init_( &lm );
+		}
+		catch ( const std::exception& e )
+		{
+			SCONE_ERROR( "Error in " + script_->GetFile().str() + " while calling init(): " + e.what() );
+		}
 
 		model.AddExternalResource( script_->GetFile() );
 	}
@@ -25,16 +34,32 @@ namespace scone
 
 	double ScriptMeasure::ComputeResult( Model& model )
 	{
-		lua_model lm( const_cast<Model&>( model ) );
-		double res = result_( &lm );
-		return res;
+		try
+		{
+			lua_model lm( const_cast<Model&>( model ) );
+			return result_( &lm );
+		}
+		catch ( const std::exception& e )
+		{
+			log::error( "Error in ", script_->GetFile().str(), " while calling result(): ", e.what() );
+			return WorstResult();
+		}
 	}
 
 	bool ScriptMeasure::UpdateMeasure( const Model& model, double timestamp )
 	{
-		lua_model lm( const_cast<Model&>( model ) );
-		double res = update_( &lm );
-		return res;
+		try
+		{
+			lua_model lm( const_cast<Model&>( model ) );
+			update_( &lm );
+		}
+		catch ( const std::exception& e )
+		{
+			log::error( "Error in ", script_->GetFile().str(), " while calling update(): ", e.what() );
+			return true;
+		}
+
+		return false;
 	}
 
 	String ScriptMeasure::GetClassSignature() const
