### 2019-05-06 caf11cf
U: added option for external parameters in scripts, cleanup + refactor
