+++ src/sconelib/sconelua/ScriptMeasure.cpp
@@ -1,7 +1,7 @@
 #include "ScriptMeasure.h"
 
-#include "LuaScript.h"
-#include "scone_lua_api.h"
+#include "lua_script.h"
+#include "lua_api.h"
 #include "scone/core/Log.h"
 
 namespace scone
@@ -9,11 +9,11 @@ namespace scone
 
 	ScriptMeasure::ScriptMeasure( const PropNode& props, Params& par, Model& model, const Location& loc ) :
 	Measure( props, par, model, loc ),
-	script_( new LuaScript( props, par, model ) )
+	script_( new lua_script( props, par, model ) )
 	{
-		init_ = script_->GetFunction( "init" );
-		update_ = script_->GetFunction( "update" );
-		result_ = script_->GetFunction( "result" );
+		init_ = script_->find_function( "init" );
+		update_ = script_->find_function( "update" );
+		result_ = script_->find_function( "result" );
 
 		lua_model lm( model );
 		try
@@ -23,10 +23,10 @@ namespace scone
 		}
 		catch ( const std::exception& e )
 		{
-			SCONE_ERROR( "Error in " + script_->GetFile().str() + " while calling init(): " + e.what() );
+			SCONE_ERROR( "Error in " + script_->script_file_.str() + " while calling init(): " + e.what() );
 		}
 
-		model.AddExternalResource( script_->GetFile() );
+		model.AddExternalResource( script_->script_file_ );
 	}
 
 	ScriptMeasure::~ScriptMeasure()
@@ -41,7 +41,7 @@ namespace scone
 		}
 		catch ( const std::exception& e )
 		{
-			log::error( "Error in ", script_->GetFile().str(), " while calling result(): ", e.what() );
+			log::error( "Error in ", script_->script_file_, " while calling result(): ", e.what() );
 			return WorstResult();
 		}
 	}
@@ -51,15 +51,13 @@ namespace scone
 		try
 		{
 			lua_model lm( const_cast<Model&>( model ) );
-			update_( &lm );
+			return update_( &lm );
 		}
 		catch ( const std::exception& e )
 		{
-			log::error( "Error in ", script_->GetFile().str(), " while calling update(): ", e.what() );
+			log::error( "Error in ", script_->script_file_, " while calling update(): ", e.what() );
 			return true;
 		}
-
-		return false;
 	}
 
 	String ScriptMeasure::GetClassSignature() const
