+++ src/sconelib/scone/model/Model.h
@@ -44,7 +44,6 @@ namespace scone
 		/// body access
 		std::vector< BodyUP >& GetBodies() { return m_Bodies; }
 		const std::vector< BodyUP >& GetBodies() const { return m_Bodies; }
-		Body* GetBody( index_t i ) { return m_Bodies[ i ].get(); }
 
 		/// joint access
 		std::vector< JointUP >& GetJoints() { return m_Joints; }
