### 2019-05-06 d847b17
N: Added stub code for performance test
