### 2019-05-07 1905979
N: Added performance test, some refactoring and cleanup
