+++ src/sconestudio/SconeStudio.cpp
@@ -73,6 +73,7 @@ scene_( true )
 	addMenuAction( scenarioMenu, "Run &Multiple Optimizations", this, &SconeStudio::optimizeScenarioMultiple, QKeySequence( "Ctrl+Shift+F5" ), true );
 	addMenuAction( scenarioMenu, "&Abort Optimizations", this, &SconeStudio::abortOptimizations, QKeySequence(), true );
 	addMenuAction( scenarioMenu, "&Run Scenario", this, &SconeStudio::runScenario, QKeySequence( "Ctrl+T" ) );
+	addMenuAction( scenarioMenu, "Measure Scenario &Performance", this, &SconeStudio::performanceTest, QKeySequence( "Ctrl+P" ) );
 
 	auto toolsMenu = menuBar()->addMenu( "&Tools" );
 	addMenuAction( toolsMenu, "Generate &Video...", this, &SconeStudio::createVideo );
@@ -176,11 +177,11 @@ void SconeStudio::runSimulation( const QString& filename )
 	if ( createModel( filename.toStdString() ) )
 	{
 		updateViewSettings();
-		storageModel.setStorage( &model->GetData() );
+		storageModel.setStorage( &model_->GetData() );
 		analysisView->reset();
-		if ( model->IsEvaluating() )
+		if ( model_->IsEvaluating() )
 			evaluate();
-		ui.playControl->setRange( 0, model->GetMaxTime() );
+		ui.playControl->setRange( 0, model_->GetMaxTime() );
 	}
 	SCONE_PROFILE_REPORT;
 }
@@ -190,7 +191,7 @@ void SconeStudio::activateBrowserItem( QModelIndex idx )
 	currentParFile = ui.resultsBrowser->fileSystemModel()->fileInfo( idx ).absoluteFilePath();
 	ui.playControl->reset();
 	runSimulation( currentParFile );
-	if ( model )
+	if ( model_ )
 		ui.playControl->play();
 }
 
@@ -218,9 +219,9 @@ void SconeStudio::refreshAnalysis()
 
 void SconeStudio::evaluate()
 {
-	SCONE_ASSERT( model );
+	SCONE_ASSERT( model_ );
 
-	QProgressDialog dlg( ( "Evaluating " + model->GetFileName().str() ).c_str(), "Abort", 0, 1000, this );
+	QProgressDialog dlg( ( "Evaluating " + model_->GetFileName().str() ).c_str(), "Abort", 0, 1000, this );
 	dlg.setWindowModality( Qt::WindowModal );
 	dlg.show();
 	QApplication::processEvents();
@@ -231,16 +232,16 @@ void SconeStudio::evaluate()
 	const xo::time visual_update = 250_ms;
 	xo::time prev_visual_time = xo::time() - visual_update;
 	xo::timer real_time;
-	for ( double t = step_size; model->IsEvaluating(); t += step_size )
+	for ( double t = step_size; model_->IsEvaluating(); t += step_size )
 	{
 		auto rt = real_time();
 		if ( rt - prev_visual_time >= visual_update )
 		{
 			// update 3D visuals and progress bar
 			setTime( t, true );
-			dlg.setValue( int( 1000 * t / model->GetMaxTime() ) );
+			dlg.setValue( int( 1000 * t / model_->GetMaxTime() ) );
 			if ( dlg.wasCanceled() ) {
-				model->FinalizeEvaluation( false );
+				model_->FinalizeEvaluation( false );
 				break;
 			}
 			prev_visual_time = rt;
@@ -250,18 +251,18 @@ void SconeStudio::evaluate()
 
 	// report duration
 	auto real_dur = real_time().seconds<double>();
-	auto sim_time = model->GetTime();
+	auto sim_time = model_->GetTime();
[…]
