### 2019-05-07 f98275e
F: Measure now defaults to class name if no name is given
