### 2019-05-08 0f3b23b
U: version bump to 1.0.0
