+++ src/sconelib/scone/core/Profiler.h
@@ -19,17 +19,17 @@
 #if defined SCONE_ENABLE_PROFILING
 #	define SCONE_PROFILE_FUNCTION ScopedProfile unique_scoped_function_profile( Profiler::GetGlobalInstance(), __FUNCTION__ )
 #	define SCONE_PROFILE_SCOPE( scope_name_arg ) ScopedProfile unique_scoped_function_profile( Profiler::GetGlobalInstance(), scope_name_arg )
-#	define SCONE_PROFILE_RESET Profiler::GetGlobalInstance().Reset()
+#	define SCONE_PROFILE_START Profiler::GetGlobalInstance().Reset()
 #	define SCONE_PROFILE_REPORT log::info( Profiler::GetGlobalInstance().GetReport() )
 #elif defined SCONE_ENABLE_XO_PROFILING
-#	define SCONE_PROFILE_FUNCTION xo::scoped_section_profiler unique_scoped_profile( __FUNCTION__ )
-#	define SCONE_PROFILE_SCOPE( scope_name_arg ) xo::scoped_section_profiler unique_scoped_profile( scope_name_arg )
-#	define SCONE_PROFILE_RESET xo::profiler::instance().reset()
+#	define SCONE_PROFILE_FUNCTION xo::scope_profiler unique_scoped_profile( __FUNCTION__, xo::profiler::instance() )
+#	define SCONE_PROFILE_SCOPE( scope_name_arg ) xo::scope_profiler unique_scoped_profile( scope_name_arg, xo::profiler::instance() )
+#	define SCONE_PROFILE_START xo::profiler::instance().start()
 #	define SCONE_PROFILE_REPORT log::info( xo::profiler::instance().report() )
 #else 
 #	define SCONE_PROFILE_FUNCTION
 #	define SCONE_PROFILE_SCOPE( scope_name_arg )
-#	define SCONE_PROFILE_RESET
+#	define SCONE_PROFILE_START
 #	define SCONE_PROFILE_REPORT PropNode()
 #endif
 
