### 2019-05-10 84210a6
U: profiler optimizations and cleanup
