+++ src/sconelib/sconeopensim3/MuscleOpenSim3.cpp
@@ -33,107 +33,92 @@ namespace scone
 		return m_osMus.getName();
 	}
 
-	scone::Real MuscleOpenSim3::GetOptimalFiberLength() const
+	Real MuscleOpenSim3::GetOptimalFiberLength() const
 	{
 		return m_osMus.getOptimalFiberLength();
 	}
 
-	scone::Real MuscleOpenSim3::GetTendonSlackLength() const
+	Real MuscleOpenSim3::GetTendonSlackLength() const
 	{
 		return m_osMus.getTendonSlackLength();
 	}
 
-	scone::Real MuscleOpenSim3::GetForce() const
+	Real MuscleOpenSim3::GetForce() const
 	{
-		SCONE_PROFILE_FUNCTION;
 		// OpenSim: why can't I just use getWorkingState()?
 		// OpenSim: why must I update to Dynamics for getForce()?
 		m_Model.GetOsimModel().getMultibodySystem().realize( m_Model.GetTkState(), SimTK::Stage::Dynamics );
 		return m_osMus.getForce( m_Model.GetTkState() );
 	}
 
-	scone::Real MuscleOpenSim3::GetNormalizedForce() const
+	Real MuscleOpenSim3::GetNormalizedForce() const
 	{
-		SCONE_PROFILE_FUNCTION;
 		return GetForce() / GetMaxIsometricForce();
 	}
 
-	scone::Real scone::MuscleOpenSim3::GetLength() const
+	Real MuscleOpenSim3::GetLength() const
 	{
-		SCONE_PROFILE_FUNCTION;
 		m_Model.GetOsimModel().getMultibodySystem().realize( m_Model.GetTkState(), SimTK::Stage::Position );
 		return m_osMus.getLength( m_Model.GetTkState() );
 	}
 
-	scone::Real MuscleOpenSim3::GetVelocity() const
+	Real MuscleOpenSim3::GetVelocity() const
 	{
-		SCONE_PROFILE_FUNCTION;
 		m_Model.GetOsimModel().getMultibodySystem().realize( m_Model.GetTkState(), SimTK::Stage::Velocity );
 		return m_osMus.getLengtheningSpeed( m_Model.GetTkState() );
 	}
 
-	scone::Real MuscleOpenSim3::GetFiberForce() const
+	Real MuscleOpenSim3::GetFiberForce() const
 	{
-		SCONE_PROFILE_FUNCTION;
 		return m_osMus.getFiberForce( m_Model.GetTkState() );
 	}
 
-	scone::Real MuscleOpenSim3::GetNormalizedFiberForce() const
+	Real MuscleOpenSim3::GetNormalizedFiberForce() const
 	{
-		SCONE_PROFILE_FUNCTION;
 		return m_osMus.getFiberForce( m_Model.GetTkState() ) / m_osMus.getMaxIsometricForce();
 	}
 
-	scone::Real MuscleOpenSim3::GetActiveFiberForce() const
+	Real MuscleOpenSim3::GetActiveFiberForce() const
 	{
-		SCONE_PROFILE_FUNCTION;
 		return m_osMus.getActiveFiberForce( m_Model.GetTkState() );
 	}
 
-	scone::Real scone::MuscleOpenSim3::GetFiberLength() const
+	Real MuscleOpenSim3::GetFiberLength() const
 	{
-		SCONE_PROFILE_FUNCTION;
 		return m_osMus.getFiberLength( m_Model.GetTkState() );
 	}
 
-	scone::Real MuscleOpenSim3::GetNormalizedFiberLength() const
+	Real MuscleOpenSim3::GetNormalizedFiberLength() const
 	{
-		SCONE_PROFILE_FUNCTION;
[…]
