+++ src/sconelib/scone/model/Leg.cpp
@@ -34,7 +34,6 @@ namespace scone
 
 	Vec3 Leg::GetContactForce() const
 	{
-		SCONE_PROFILE_FUNCTION;
 		return m_Foot.GetBody().GetContactForce();
 	}
 
@@ -50,7 +49,6 @@ namespace scone
 
 	void Leg::GetContactForceMomentCop( Vec3& force, Vec3& moment, Vec3& cop ) const
 	{
-		SCONE_PROFILE_FUNCTION;
 		force = m_Foot.GetBody().GetContactForce();
 		moment = m_Foot.GetBody().GetContactMoment();
 		cop = GetGroundCop( force, moment );
