### 2019-05-10 723a420
R: moved SCONE_PROFILE_* macros from Profiler.h to profiler_config.h
