+++ src/sconestudio/StudioModel.cpp
@@ -10,7 +10,7 @@
 
 #include "scone/optimization/opt_tools.h"
 #include "scone/core/StorageIo.h"
-#include "scone/core/Profiler.h"
+#include "scone/core/profiler_config.h"
 #include "scone/core/Factories.h"
 #include "scone/core/system_tools.h"
 #include "scone/core/Settings.h"
