+++ src/sconelib/sconelua/ScriptController.cpp
@@ -13,7 +13,7 @@
 #include "lua_script.h"
 #include "lua_api.h"
 #include "scone/core/Log.h"
-#include "scone/core/Profiler.h"
+#include "scone/core/profiler_config.h"
 
 namespace scone
 {
