### 2019-05-10 d8d8ea1
U: profiling is now enabled through CMake instead of platform.h
