+++ src/sconelib/sconeopensim3/CMakeLists.txt
@@ -32,7 +32,7 @@ set_target_properties(sconeopensim3 PROPERTIES CXX_STANDARD 17 CXX_STANDARD_REQU
 
 if (MSVC)
 	target_compile_definitions( sconeopensim3 PRIVATE SCONE_MODEL_OPENSIM_3_EXPORTS )
-	target_compile_definitions( sconeopensim3 PRIVATE "$<$<CONFIG:RELWITHDEBINFO>:SCONE_ENABLE_XO_PROFILING>" )
+	target_compile_definitions( sconeopensim3 PRIVATE SCONE_ENABLE_XO_PROFILING )
 endif()
 
 source_group("" FILES ${MODEL_OPENSIM_3_FILES})
