+++ src/sconelib/scone/optimization/CmaOptimizerSpot.h
@@ -11,6 +11,7 @@
 #include "CmaOptimizer.h"
 #include "spot/cma_optimizer.h"
 #include "spot/reporter.h"
+#include "spot/file_reporter.h"
 #include "xo/system/log_sink.h"
 
 namespace scone
