### 2019-05-14 65eb5ba
F: sconeunittest output is now send to temp folder
