### 2019-05-14 80ec623
U: updated error message in SetProperties
