### 2019-05-14 51c29cd
U: updated error message in SetProperties
