### 2019-05-15 437ae25
U: Removed FindXo / FindSpot, now simply adds submodule root folder. Incorporated XO_TEST_ENABLED.
