### 2019-05-16 293f972
F*: JumpMeasure now supports both 'body' and 'target_body', matching documentation and keeping backwards compatibility
