### 2019-05-17 efc7dce
N*: ScriptMeasure and ScriptController now support store_data() for writing data to sto (and SCONE analysis window) U*: Improved error handling and reporting for scripts
