+++ src/sconelib/sconelua/ScriptMeasure.h
@@ -15,9 +15,9 @@ namespace scone
 		
 		virtual double ComputeResult( Model& model ) override;
 		virtual bool UpdateMeasure( const Model& model, double timestamp ) override;
+		virtual void StoreData( Storage<Real>::Frame& frame, const StoreDataFlags& flags ) const override;
 
 	protected:
-
 		virtual String GetClassSignature() const override;
 
 	private:
@@ -25,5 +25,6 @@ namespace scone
 		std::function<void( struct lua_model* )> init_;
 		std::function<bool( struct lua_model* )> update_;
 		std::function<double( struct lua_model* )> result_;
+		std::function<double( struct lua_frame* )> store_;
 	};
 }
