### 2019-05-17 cefc4c2
U: simulation_test now includes computer name in results
