### 2019-05-20 bc91f65
N*: Added support for Find (Ctrl+F) and Find Next / Previous (F3 / Shift+F4) in text editor F*: Text modification detection now works properly with undo F*: Text document now automatically receives focus when opened
