+++ src/sconestudio/SconeStudio.h
@@ -76,6 +76,9 @@ public slots:
 	void showSettingsDialog() { settings.showDialog( this ); }
 	void setPlaybackTime( TimeInSeconds t ) { setTime( t, true ); }
 	void updateTabTitles();
+	void findDialog() { if ( auto* e = getActiveScenario() ) e->findDialog(); }
+	void findNext() { if ( auto* e = getActiveScenario() ) e->findNext(); }
+	void findPrevious() { if ( auto* e = getActiveScenario() ) e->findNext( true ); }
 
 public:
 	bool close_all;
