### 2019-05-20 d8db496
N: new tutorial_test, cleanup
