+++ src/sconelib/scone/core/version.cpp
@@ -14,7 +14,7 @@
 constexpr int SCONE_VERSION_MAJOR = 1;
 constexpr int SCONE_VERSION_MINOR = 1;
 constexpr int SCONE_VERSION_PATCH = 0;
-constexpr const char* SCONE_VERSION_POSTFIX = "BETA 2";
+constexpr const char* SCONE_VERSION_POSTFIX = "RELEASE CANDIDATE 1";
 
 namespace scone
 {
