### 2019-05-20 557995e
U*: lua scripting added body:point_pos and body:point_vel
