### 2019-05-20 #1421 @modenaxe
@tgeijten I installed SCONE in the past and used it, but now, after I uninstalled it completely and reinstalled v1.0.0, I cannot run it after installation. I am prompted the following error:

![image](https://user-images.githubusercontent.com/7674529/58027397-74c0db00-7b10-11e9-9749-0312babc30cc.png)

I have already tried uninstalling again, removing all installation files and resources, restarting the machine and reinstalling but it doesn't seem to work, same error that I have reported above. I have already check if the issue comes from the antivirus blocking the executable, but it doesn't seem so. Any idea on what could be the cause and how to fix it? My computer runs Windows 10 Pro (64-bit), let me know if you need more details of my system.
