### 2019-05-20 #1421 @tgeijten
@modenaxe sorry for the trouble! It seems you have an old configuration file that's no longer valid. You can delete to following folder to reset the configuration:

`C:\Users\<UserName>\AppData\Local\SCONE`

After that, everything should work fine.
