### 2019-05-20 #1421 @modenaxe
First of all, I want to congratulate @tgeijten for the quality of this software. Graphical interface and visualisation window in particular are truly remarkable for a research tool. I think that SCONE is an ambitious project with huge potential and creating the conditions for interested researchers of learning how to use it will be of paramount importance for its long-term success.
### Software paper
#### A statement of need
While the motivation for SCONE is well explained, and numerous examples of use are mentioned, the target audience seems very broad in scope, i.e. the paper mentions researchers interested in biological movement. As explained in more details in the Documentation comments, the SCONE user is assumed to be familiar with advanced concepts in order to build his/her own scenarios, and this already stratifies users, in a sense. I think that the target audience could be better identified specifying their intended level of computational skills, for example. Multiple typologies of users could also be identified, e.g. researchers in the field of computational biomechanics and neuromechanics, interested in developing new SCONE scenarios, controllers etc. and other researchers focusing on more clinical applications, e.g. using existing scenarios to investigate interventions outcomes by modifying just the model etc.
#### Functionality
“The ability to simultaneously optimize both model and control parameters” is listed as one of the features of SCONE. What aspects of the model can be optimised? Muscle parameters or also the musculoskeletal geometry? I would list some example.
What is the "thin API layer" needed to connect SCONE to a musculoskeletal simulation package doing? Implementing methods to read in models and input data from a specific specific format? I would add few words about this feature.
### Documentation
#### A statement of need
The same comment about the software paper applies to the documentation, since the description of the software in the SCONE website is identical the submitted paper.
#### Functionality documentation
Considering the broad audience of the project, the users will have variable background in terms of computer skills. The documentation is currently minimal, probably sufficient for a developer but surely not enough for a clinician or someone not belonging to a computational field of research. Also, the concepts of "model, controller, objective, and optimization strategy" are given for granted in the [documentation](https://scone.software/doku.php?id=doc:start). I would include in the documentation appropriate links and references to help the users to gather at least some basic information about these [main conceptual components](https://scone.software/doku.php?id=doc:concepts) of SCONE and specify how the documentation will grow in the future (if there is a plan in this respect).
Also, in my opinion, few additional points could be added in the current documentation:
1. It is not explained how to stop the simulations before the optimization is complete. Is it enough to use “Abort Optimizations”? Can the optimizations of the scenario be restarted from that point at another time?
2. Can any OpenSim model be used in a scenario or just 2D models? Is there any guideline on how to build scenarios with new models?
3. In my machine, the "Optimization Results" window does not update while I am optimizing the scenario, when the optimization is concluded or when I abort the optimization. It updates only if I restart the application, is this a bug or a feature to document?

Minor note: In the first tutorial, "File>open scenario" does not correspond to the software, where there is only “open”
#### Community guidelines
It seems that feature requests and bug tracking at the moment are done via github and not via the the simtk.org links provide on the SCONE website. Maybe it is worth including the github links together with simtk?
