### 2019-05-21 #1421 @tgeijten
Thank you @modenaxe for the great feedback. I will add the suggested content to the documentation in the upcoming days. It is indeed our plan to develop a user base outside core developers, and feedback like this is essential to achieve that.

In general, the plan is to continuously improve the website and documentation based on user feedback. Funding has been secured to enable this until at least the end of 2020; we are currently investigating various options to extend funding after that.

The link to the issue tracker and the tutorial inconsistency have already been fixed.

Concerning your issue of the Optimization Results not updating, this is indeed a software issue (one I have not encountered before). It would be highly appreciated if you can [report the issue](https://github.com/opensim-org/SCONE) so we can tackle it properly.
