### 2019-05-21 0905999
N*: Added option 'use_peek' in RangePenalty
