### 2019-05-21 13fcf76
N*: Added new MuscleMeasure for penalizing muscle activation, length and contraction velocity (issue #36).
