### 2019-05-22 #1421 @modenaxe
you're welcome @tgeijten! I realised I have another comment about functionality. Can SCONE be used also with just torque actuators acting at the joints? That would make the software of interest also for robotics. (This is more of a personal curiosity than a comment related to the software manuscript.)
