### 2019-05-22 #1421 @tgeijten
Yes -- that's a good point. It can work with any actuator, although there is somewhat of a focus on neuromuscular control strategies. But it can definitely be used for robotics, especially assistive devices. I'll adapt the text to make that more clear.
