### 2019-05-28 24c4b9f
U*: RangePenalty now supports 'mode' setting, allowing the final penalty to be the 'average' (default), 'lowest', or 'highest' of all recorded penalty values.
