### 2019-05-29 bd33b3a
F*: NoiseController more accurately handles cases where the noise STD is zero
