+++ src/sconelib/scone/controllers/NoiseController.cpp
@@ -13,23 +13,22 @@
 
 namespace scone
 {
-	
 	NoiseController::NoiseController( const PropNode& props, Params& par, Model& model, const Location& loc ) :
 	Controller( props, par, model, loc ),
 	random_seed( props.get< unsigned int >( "random_seed", 123 ) ),
 	rng_( random_seed )
 	{
-		INIT_PROP( props, base_noise,  );
+		INIT_PROP( props, base_noise, 0 );
 		INIT_PROP( props, proportional_noise, 0 );
 	}
 
 	bool NoiseController::ComputeControls( Model& model, double timestamp )
 	{
-
 		for ( auto& a : model.GetActuators() )
 		{
-			auto input = std::max( a->GetInput(), 1e-6 ); // input must be > 0
-			a->AddInput( rng_.norm( 0.0, base_noise + proportional_noise * input ) );
+			auto noise_std = base_noise + proportional_noise * a->GetInput();
+			if ( noise_std > 0.0 )
+				a->AddInput( rng_.norm( 0.0, noise_std ) );
 		}
 		return false;
 	}
