### 2019-05-29 2c18fce
F*: SequentialController now works correctly
