### 2019-05-29 355afe4
U*: CompositeController displays warning when child controllers have identical names
