### 2019-05-29 e1983b7
N*: Controllers now have 'name' member that is used as a prefix for all parameter names. This should prevent undesired parameter name clashes when combining controllers.
