+++ src/sconelib/scone/controllers/Controller.cpp
@@ -12,11 +12,12 @@
 namespace scone
 {
 	Controller::Controller( const PropNode& props, Params& par, Model& model, const Location& target_area ) :
-	HasSignature( props )
+	HasSignature( props ),
+	INIT_MEMBER( props, name, "" )
 	{
+		INIT_PROP( props, disabled_, false );
 		INIT_PAR( props, par, start_time, 0.0 );
 		INIT_PAR( props, par, stop_time, 0.0 );
-		INIT_PROP( props, disabled_, false );
 	}
 
 	Controller::~Controller()
