+++ src/sconelib/scone/core/version.cpp
@@ -13,8 +13,8 @@
 
 constexpr int SCONE_VERSION_MAJOR = 1;
 constexpr int SCONE_VERSION_MINOR = 1;
-constexpr int SCONE_VERSION_PATCH = 0;
-constexpr const char* SCONE_VERSION_POSTFIX = "RELEASE CANDIDATE 1";
+constexpr int SCONE_VERSION_PATCH = 1;
+constexpr const char* SCONE_VERSION_POSTFIX = "";
 
 namespace scone
 {
