### 2019-05-29 a49c768
R: Changed casing in lua_* classes to Lua*
