+++ src/sconelib/sconelua/lua_api.h
@@ -28,7 +28,7 @@ namespace scone
 	}
 
 	/// access to scone logging in script
-	struct lua_scone
+	struct LuaScone
 	{
 		/// display trace message
 		static void trace( const std::string& msg ) { log::trace( msg ); }
@@ -43,15 +43,15 @@ namespace scone
 	};
 
 	/// 3d vector type with components x, y, z
-	struct lua_vec3 : public Vec3d
+	struct LuaVec3 : public Vec3d
 	{
 		using Vec3::vec3_;
 	};
 
 	/// access to writing data for scone Analysis window
-	struct lua_frame
+	struct LuaFrame
 	{
-		lua_frame( Storage<Real>::Frame& f ) : frame_( f ) {}
+		LuaFrame( Storage<Real>::Frame& f ) : frame_( f ) {}
 
 		/// set a numeric value for channel named key
 		void set_value( const std::string& key, double value ) { frame_[ key ] = value; }
@@ -65,9 +65,9 @@ namespace scone
 
 	/// actuator type for use in lua scripting.
 	/// See ScriptController and ScriptMeasure for details on scripting.
-	struct lua_actuator
+	struct LuaActuator
 	{
-		lua_actuator( Actuator& a ) : act_( a ) {}
+		LuaActuator( Actuator& a ) : act_( a ) {}
 
 		/// get the name of the actuator
 		const char* name() { return act_.GetName().c_str(); }
@@ -81,9 +81,9 @@ namespace scone
 
 	/// dof (degree-of-freedom) type for use in lua scripting.
 	/// See ScriptController and ScriptMeasure for details on scripting.
-	struct lua_dof
+	struct LuaDof
 	{
-		lua_dof( Dof& d ) : dof_( d ) {}
+		LuaDof( Dof& d ) : dof_( d ) {}
 
 		/// get the name of the muscle
 		const char* name() { return dof_.GetName().c_str(); }
@@ -97,9 +97,9 @@ namespace scone
 
 	/// muscle type for use in lua scripting.
 	/// See ScriptController and ScriptMeasure for details on scripting.
-	struct lua_muscle
+	struct LuaMuscle
 	{
-		lua_muscle( Muscle& m ) : mus_( m ) {}
+		LuaMuscle( Muscle& m ) : mus_( m ) {}
 
 		/// get the name of the muscle
 		const char* name() { return mus_.GetName().c_str(); }
@@ -133,24 +133,24 @@ namespace scone
 
 	/// body type for use in lua scripting.
 	/// See ScriptController and ScriptMeasure for details on scripting.
-	struct lua_body
+	struct LuaBody
 	{
-		lua_body( Body& b ) : bod_( b ) {}
+		LuaBody( Body& b ) : bod_( b ) {}
 
 		/// get the name of the body
 		const char* name() { return bod_.GetName().c_str(); }
 		/// get the current com position [m]
-		lua_vec3 com_pos() { return bod_.GetComPos(); }
+		LuaVec3 com_pos() { return bod_.GetComPos(); }
 		/// get the current com velocity [m/s]
-		lua_vec3 com_vel() { return bod_.GetComVel(); }
+		LuaVec3 com_vel() { return bod_.GetComVel(); }
 		/// get the global position [m] of a local point p on the body
-		lua_vec3 point_pos( const lua_vec3& p ) { return bod_.GetPosOfPointOnBody( p ); }
+		LuaVec3 point_pos( const LuaVec3& p ) { return bod_.GetPosOfPointOnBody( p ); }
 		/// get the global linear velocity [m/s] of a local point p on the body
[…]
