+++ src/sconelib/sconelua/lua_api.cpp
@@ -5,86 +5,86 @@ namespace scone
 {
 	void register_lua_wrappers( sol::state& lua )
 	{
-		lua.new_usertype<lua_vec3>( "lua_vec3", sol::constructors<Vec3( double, double, double )>(),
-			"x", &lua_vec3::x,
-			"y", &lua_vec3::y,
-			"z", &lua_vec3::z
+		lua.new_usertype<LuaVec3>( "lua_vec3", sol::constructors<Vec3( double, double, double )>(),
+			"x", &LuaVec3::x,
+			"y", &LuaVec3::y,
+			"z", &LuaVec3::z
 			);
 
-		lua.new_usertype<lua_frame>( "lua_frame", sol::constructors<>(),
-			"set_value", &lua_frame::set_value,
-			"set_bool", &lua_frame::set_bool,
-			"time", &lua_frame::time
+		lua.new_usertype<LuaFrame>( "lua_frame", sol::constructors<>(),
+			"set_value", &LuaFrame::set_value,
+			"set_bool", &LuaFrame::set_bool,
+			"time", &LuaFrame::time
 			);
 
-		lua.new_usertype<lua_actuator>( "lua_actuator", sol::constructors<>(),
-			"name", &lua_actuator::name,
-			"add_input", &lua_actuator::add_input,
-			"input", &lua_actuator::input
+		lua.new_usertype<LuaActuator>( "lua_actuator", sol::constructors<>(),
+			"name", &LuaActuator::name,
+			"add_input", &LuaActuator::add_input,
+			"input", &LuaActuator::input
 			);
 
-		lua.new_usertype<lua_dof>( "lua_dof", sol::constructors<>(),
-			"name", &lua_dof::name,
-			"position", &lua_dof::position,
-			"velocity", &lua_dof::velocity
+		lua.new_usertype<LuaDof>( "lua_dof", sol::constructors<>(),
+			"name", &LuaDof::name,
+			"position", &LuaDof::position,
+			"velocity", &LuaDof::velocity
 			);
 
-		lua.new_usertype<lua_body>( "lua_body", sol::constructors<>(),
-			"name", &lua_body::name,
-			"com_pos", &lua_body::com_pos,
-			"com_vel", &lua_body::com_vel,
-			"ang_pos", &lua_body::ang_pos,
-			"ang_vel", &lua_body::ang_vel,
-			"add_external_force", &lua_body::add_external_force,
-			"add_external_moment", &lua_body::add_external_moment
+		lua.new_usertype<LuaBody>( "lua_body", sol::constructors<>(),
+			"name", &LuaBody::name,
+			"com_pos", &LuaBody::com_pos,
+			"com_vel", &LuaBody::com_vel,
+			"ang_pos", &LuaBody::ang_pos,
+			"ang_vel", &LuaBody::ang_vel,
+			"add_external_force", &LuaBody::add_external_force,
+			"add_external_moment", &LuaBody::add_external_moment
 			);
 
-		lua.new_usertype<lua_muscle>( "lua_muscle", sol::constructors<>(),
-			"name", &lua_muscle::name,
-			"add_input", &lua_muscle::add_input,
-			"input", &lua_muscle::input,
-			"excitation", &lua_muscle::excitation,
-			"activation", &lua_muscle::activation,
-			"fiber_length", &lua_muscle::fiber_length,
-			"normalized_fiber_length", &lua_muscle::normalized_fiber_length,
-			"optimal_fiber_length", &lua_muscle::optimal_fiber_length,
-			"fiber_force", &lua_muscle::fiber_force,
-			"normalized_fiber_force", &lua_muscle::normalized_fiber_force,
-			"max_isometric_force", &lua_muscle::max_isometric_force,
[…]
