+++ src/sconelib/sconelua/ScriptMeasure.h
@@ -7,6 +7,8 @@
 
 namespace scone
 {
+	/// Measure defined through a Lua script.
+	/// Documentation is in preparation; see Tutorial 6a and 6b for example usage.
 	class SCONE_LUA_API ScriptMeasure : public Measure
 	{
 	public:
