### 2019-05-29 #1421 @tgeijten
@Kevin-Mattheus-Moerman do I need finish all suggested documentation updates first or can we proceed with the publication?
