### 2019-06-04 2d53e81
U*: Updated reference documentation for lua_api, removed Doxygen-style function documentation for Model and Controller.
