+++ src/sconelib/sconelua/lua_api.h
@@ -16,6 +16,8 @@ namespace sol { class state; }
 
 namespace scone
 {
+	using LuaString = const char*;
+
 	template< typename T > T& GetByLuaIndex( std::vector<T>& vec, int index ) {
 		SCONE_ERROR_IF( index < 1 || index > vec.size(), "Index must be between 1 and " + xo::to_str( vec.size() ) );
 		return vec[ index - 1 ];
@@ -27,19 +29,19 @@ namespace scone
 		return *it;
 	}
 
-	/// access to scone logging in script
+	/// Access to scone logging in script
 	struct LuaScone
 	{
 		/// display trace message
-		static void trace( const std::string& msg ) { log::trace( msg ); }
+		static void trace( LuaString msg ) { log::trace( msg ); }
 		/// display debug message
-		static void debug( const std::string& msg ) { log::debug( msg ); }
+		static void debug( LuaString msg ) { log::debug( msg ); }
 		/// display info message
-		static void info( const std::string& msg ) { log::info( msg ); }
+		static void info( LuaString msg ) { log::info( msg ); }
 		/// display warning message
-		static void warning( const std::string& msg ) { log::warning( msg ); }
+		static void warning( LuaString msg ) { log::warning( msg ); }
 		/// display error message
-		static void error( const std::string& msg ) { log::error( msg ); }
+		static void error( LuaString msg ) { log::error( msg ); }
 	};
 
 	/// 3d vector type with components x, y, z
@@ -48,29 +50,29 @@ namespace scone
 		using Vec3::vec3_;
 	};
 
-	/// access to writing data for scone Analysis window
+	/// Access to writing data for scone Analysis window
 	struct LuaFrame
 	{
 		LuaFrame( Storage<Real>::Frame& f ) : frame_( f ) {}
 
 		/// set a numeric value for channel named key
-		void set_value( const std::string& key, double value ) { frame_[ key ] = value; }
+		void set_value( LuaString key, double value ) { frame_[ key ] = value; }
 		/// set a boolean (true or false) value for channel named key
-		void set_bool( const std::string& key, bool b ) { frame_[ key ] = b ? 1.0 : 0.0; }
+		void set_bool( LuaString key, bool b ) { frame_[ key ] = b ? 1.0 : 0.0; }
 		/// get time of current frame
 		double time() { return frame_.GetTime(); }
 
 		Storage<Real>::Frame& frame_;
 	};
 
-	/// actuator type for use in lua scripting.
+	/// Actuator type for use in lua scripting.
 	/// See ScriptController and ScriptMeasure for details on scripting.
 	struct LuaActuator
 	{
 		LuaActuator( Actuator& a ) : act_( a ) {}
 
 		/// get the name of the actuator
-		const char* name() { return act_.GetName().c_str(); }
+		LuaString name() { return act_.GetName().c_str(); }
 		/// add a value to the normalized actuator input
 		void add_input( double value ) { act_.AddInput( value ); }
 		/// get the current actuator input
@@ -79,30 +81,30 @@ namespace scone
 		Actuator& act_;
 	};
 
-	/// dof (degree-of-freedom) type for use in lua scripting.
+	/// Dof (degree-of-freedom) type for use in lua scripting.
 	/// See ScriptController and ScriptMeasure for details on scripting.
 	struct LuaDof
 	{
 		LuaDof( Dof& d ) : dof_( d ) {}
 
 		/// get the name of the muscle
-		const char* name() { return dof_.GetName().c_str(); }
+		LuaString name() { return dof_.GetName().c_str(); }
 		/// get the current value (position) of the dof
-		double position() const { return dof_.GetPos(); }
[…]
