+++ src/sconelib/sconelua/lua_api.h
@@ -17,6 +17,7 @@ namespace sol { class state; }
 namespace scone
 {
 	using LuaString = const char*;
+	using LuaNumber = double;
 
 	template< typename T > T& GetByLuaIndex( std::vector<T>& vec, int index ) {
 		SCONE_ERROR_IF( index < 1 || index > vec.size(), "Index must be between 1 and " + xo::to_str( vec.size() ) );
@@ -56,11 +57,11 @@ namespace scone
 		LuaFrame( Storage<Real>::Frame& f ) : frame_( f ) {}
 
 		/// set a numeric value for channel named key
-		void set_value( LuaString key, double value ) { frame_[ key ] = value; }
+		void set_value( LuaString key, LuaNumber value ) { frame_[ key ] = value; }
 		/// set a boolean (true or false) value for channel named key
 		void set_bool( LuaString key, bool b ) { frame_[ key ] = b ? 1.0 : 0.0; }
 		/// get time of current frame
-		double time() { return frame_.GetTime(); }
+		LuaNumber time() { return frame_.GetTime(); }
 
 		Storage<Real>::Frame& frame_;
 	};
@@ -73,10 +74,10 @@ namespace scone
 
 		/// get the name of the actuator
 		LuaString name() { return act_.GetName().c_str(); }
-		/// add a value to the normalized actuator input
-		void add_input( double value ) { act_.AddInput( value ); }
-		/// get the current actuator input
-		double input() { return act_.GetInput(); }
+		/// add a value to the normalized actuator input [0..1]
+		void add_input( LuaNumber value ) { act_.AddInput( value ); }
+		/// get the current actuator input [0..1]
+		LuaNumber input() { return act_.GetInput(); }
 
 		Actuator& act_;
 	};
@@ -89,10 +90,10 @@ namespace scone
 
 		/// get the name of the muscle
 		LuaString name() { return dof_.GetName().c_str(); }
-		/// get the current value (position) of the dof
-		double position() { return dof_.GetPos(); }
-		/// get the current velocity of the dof [/s]
-		double velocity() { return dof_.GetVel(); }
+		/// get the current value (position) of the dof in [m] or [rad]
+		LuaNumber position() { return dof_.GetPos(); }
+		/// get the current velocity of the dof in [m/s] or [rad/s]
+		LuaNumber velocity() { return dof_.GetVel(); }
 
 		Dof& dof_;
 	};
@@ -106,29 +107,29 @@ namespace scone
 		/// get the name of the muscle
 		LuaString name() { return mus_.GetName().c_str(); }
 		/// add a value to the normalized actuator input
-		void add_input( double value ) { mus_.AddInput( value ); }
+		void add_input( LuaNumber value ) { mus_.AddInput( value ); }
 		/// get the current actuator input
-		double input() { return mus_.GetInput(); }
+		LuaNumber input() { return mus_.GetInput(); }
 		/// get the normalized excitation level [0..1] of the muscle
-		double excitation() { return mus_.GetExcitation(); }
+		LuaNumber excitation() { return mus_.GetExcitation(); }
 		/// get the normalized activation level [0..1] of the muscle
-		double activation() { return mus_.GetActivation(); }
+		LuaNumber activation() { return mus_.GetActivation(); }
 		/// get the fiber length [m] of the contractile element
-		double fiber_length() { return mus_.GetFiberLength(); }
+		LuaNumber fiber_length() { return mus_.GetFiberLength(); }
 		/// get the normalized fiber length of the contractile element
[…]
