### 2019-06-04 #1421 @tgeijten
@Kevin-Mattheus-Moerman thanks for the suggestion. @modenaxe, I have implemented all the changes suggested by you on the SCONE website -- I'm curious to hear if you agree. @demotu I am also eagerly awaiting your feedback!
