### 2019-06-04 d7d6fa9
R: moved xo/diagnose -> xo/system
