### 2019-06-05 858a2ef
R: Added MirrorController.cpp
