### 2019-06-05 07edb04
F*: Close current tab (Ctrl+W) now works correctly
