### 2019-06-05 86475bc
F: Added MirrorController.cpp to CMakeLists.txt
