### 2019-06-05 1524ecb
R*: The 'symmetric' property now works as expected in all controllers. If unspecified, the property is inherited from the parent controller, or set to true if the controller has no parent (i.e. the root controller).
