+++ src/sconelib/scone/model/Location.cpp
@@ -17,9 +17,9 @@ namespace scone
 		auto name_no_side = GetNameNoSide( name );
 	
 		if ( name_side == NoSide ) // no side in name, add location side
-			return name_no_side + GetSideName( side );
+			return name_no_side + GetSideName( side_ );
 		else if ( name_side == OppositeSide )
-			return name_no_side + GetSideName( GetOppositeSide( side ) );
+			return name_no_side + GetSideName( GetOppositeSide( side_ ) );
 		else return name; // keep original side
 	}
 }
