### 2019-06-05 50e4885
F: from_prop_node(): moved temp value out of if statement to avoid gcc compilation error #59
