### 2019-06-05 413cba7
R: using ptr() instead of &*
