### 2019-06-07 #1421 @Kevin-Mattheus-Moerman
@modenaxe thanks again for the awesome review efforts here! Can you review the changes implemented by @tgeijten?
