### 2019-06-10 #1421 @modenaxe
Thanks @tgeijten, I am happy with the changes. In my opinion there are these few additional points to clarify to conclude the review process:
- target users have been specified in the website but not in the paper. According to the guidelines above (the only box I haven't ticked yet) I think you should add them also in the software paper.
- in the paper "employ" should be "employed" on fourth paragraph.
- I couldn't find a comment about the use of 3D models in the website. I would include it in the FAQ section of the website.
- Can the optimization of a scenario be resumed if aborted before convergence?
- the issue tracker links are still redirecting to simtk, not github, in [License and contributing](https://scone.software/doku.php?id=license).

Other than this, I am ok with the rest, well done!
