### 2019-06-10 #1421 @Kevin-Mattheus-Moerman
@tgeijten do you currently have community guidelines?

- [ ] Can you add a `CONTRIBUTING.md` file to your repository and link to it e.g. with the heading **Contributing** in your README? There are templates you can find ([this](https://github.com/nayafia/contributing-template/blob/master/CONTRIBUTING-template.md) is a detailed one, [one of mine](https://github.com/gibbonCode/GIBBON/blob/master/CONTRIBUTING.md) is a bit shorter as an example).

- [ ] I recommend that you also add a `CODE_OF_CONDUCT.md` file and link to it in the README. A great template you can copy is available [here](https://www.contributor-covenant.org/) (be sure to enter you contact details in the COC document).
