### 2019-06-11 d8e8fab
R: depricated -> deprecated :-)
