+++ src/sconelib/scone/core/depricated/Vec3.cpp
@@ -1,204 +1,0 @@
-#include "stdafx.h"
-
-#if 0
-#include "Vec3.h"
-#include "Exception.h"
-#include <sstream>
-#include <cmath>
-#include "tools.h"
-
-namespace scone
-{
-	const Vec3 Vec3::ZERO = Vec3( 0, 0, 0 );
-	const Vec3 Vec3::UNIT_X = Vec3( 1, 0, 0 );
-	const Vec3 Vec3::UNIT_Y = Vec3( 0, 1, 0 );
-	const Vec3 Vec3::UNIT_Z = Vec3( 0, 0, 1 );
-	const Vec3 Vec3::NEG_UNIT_X = Vec3( -1, 0, 0 );
-	const Vec3 Vec3::NEG_UNIT_Y = Vec3( 0, -1, 0 );
-	const Vec3 Vec3::NEG_UNIT_Z = Vec3( 0, 0, -1 );
-
-	bool Vec3::IsZeroLength()
-	{
-		return GetSquaredLength() <= std::numeric_limits<Real>::epsilon();
-	}
-
-	Real Vec3::GetSquaredLength() const
-	{
-		return X() * X() + Y() * Y() + Z() * Z();
-	}
-
-	Real Vec3::GetLength() const
-	{
-		return sqrt( X() * X() + Y() * Y() + Z() * Z() );
-	}
-
-	Vec3 Vec3::GetNormalized() const
-	{
-		Real s = GetLength();
-		if ( s > std::numeric_limits<Real>::epsilon() )
-			return ( *this ) * ( ( Real )1.0 / s );
-		else return ( *this );
-	}
-
-	bool Vec3::Normalize()
-	{
-		Real s = GetLength();
-		if ( s <= std::numeric_limits<Real>::epsilon() )
-			return false;
-		( *this ) *= ( ( Real )1.0 / s );
-		return true;
-	}
-
-	Vec3 operator*( Real s, const Vec3& vec )
-	{
-		return Vec3( vec.X() * s, vec.Y() * s, vec.Z() * s );
-	}
-
-	Vec3 Vec3::operator*( const Vec3& vec ) const
-	{
-		return Vec3( X() * vec.X(), Y() * vec.Y(), Z() * vec.Z() );
-	}
-
-	Vec3 Vec3::operator*( Real s ) const
-	{
-		return Vec3( X() * s, Y() * s, Z() * s );
-	}
-
-	Vec3& Vec3::operator*=( const Vec3& v )
-	{
-		X() *= v.X();
-		Y() *= v.Y();
-		Z() *= v.Z();
-		return ( *this );
-	}
-
-	Vec3& Vec3::operator*=( Real s )
-	{
-		X() *= s;
-		Y() *= s;
-		Z() *= s;
-		return ( *this );
-	}
-
-	Vec3& Vec3::operator/=( Real s )
-	{
-		SCONE_ASSERT( s != Real( 0 ) );
-		s = Real( 1 ) / s;
-
-		X() *= s;
-		Y() *= s;
-		Z() *= s;
-		return ( *this );
-	}
-
-	Vec3 Vec3::operator/( Real s ) const
-	{
-		SCONE_ASSERT( s != Real( 0 ) );
-		s = Real( 1 ) / s;
-		return Vec3( X() * s, Y() * s, Z() * s );
-	}
-
-	Vec3 Vec3::operator+( const Vec3& v ) const
-	{
-		return Vec3( X() + v.X(), Y() + v.Y(), Z() + v.Z() );
-	}
-
-	Vec3& Vec3::operator+=( const Vec3& v )
-	{
-		X() += v.X();
-		Y() += v.Y();
-		Z() += v.Z();
-		return ( *this );
-	}
-
-	Vec3 Vec3::operator-( const Vec3& v ) const
[…]
