### 2019-06-12 1ac290c
R: removed SFINAE-based enum <-> string/prop_node conversion, should fix #59
