+++ src/sconelib/scone/measures/JumpMeasure.cpp
@@ -10,7 +10,7 @@
 #include "scone/model/Location.h"
 #include "scone/core/Log.h"
 #include "xo/numerical/math.h"
-#include "xo/string/string_cast_enum.h"
+#include "xo/container/prop_node_tools.h"
 
 namespace scone
 {
@@ -26,7 +26,7 @@ namespace scone
 		INIT_PROP( props, recover_time, 1.0 );
 		INIT_PROP( props, terminate_on_peak, true );
 		INIT_PROP( props, negate_result, false );
-		INIT_PROP( props, jump_type, HighJump );
+		jump_type = static_cast<JumpType>( props.get<int>( "jump_type" ) );
 		INIT_PROP( props, minimize, false ); // defaults to false
 		INIT_PROP( props, offset, Vec3::zero() );
 
