### 2019-06-12 dad706c
U: paper.md, added potential users, removed feature list.
