### 2019-06-12 9764d7b
N: Added CONTRIBUTING.md, updated README.md.
