+++ README.md
@@ -11,5 +11,8 @@ See [BUILD.md](BUILD.md) for instructions on how to build SCONE from source.
 ## Running SCONE
 Once you have SCONE installed on your system, be sure to follow the [SCONE tutorials and examples](https://scone.software/doku.php?id=tutorials:start) to get started!
 
+## Contributing to SCONE
+See [CONTRIBUTING.md](CONTRIBUTING.md) for information on ways to contribute to SCONE.
+
 ## License
 SCONE is licensed under the [GNU Public License 3.0](https://www.gnu.org/licenses/gpl-3.0.en.html)
