+++ CONTRIBUTING.md
@@ -1,0 +1,41 @@
+# Contribution Guidelines
+
+Thank you for your interest in contributing to SCONE -- your help is appreciated!
+
+Ways in which to contribute are:
+
+  * Help improve SCONE by **reporting issues** or **suggesting new features** on the [SCONE GitHub issues page](https://github.com/opensim-org/SCONE/issues)..
+  * Become involved by **posting** and **replying to posts** on the [SCONE User Forum](https://simtk.org/plugins/phpBB/indexPhpbb.php?group_id=1180&pluginname=phpBB).
+  * C++ developers can help improve the source code by sending **pull requests** (see below for more information).
+
+## Reporting issues
+
+You are encouraged to report issues on the [SCONE GitHub issues page](https://github.com/opensim-org/SCONE/issues).
+
+- **Search for existing issues.** Please check to see if someone else has reported the same issue. If so, please add any further information here.
+- **Share as much information as possible.** Include operating system and SCONE version number. Also, include steps to reproduce the bug.
+
+## Suggesting new features
+
+Feature requests and enhancements can be reported on the [SCONE GitHub issues page](https://github.com/opensim-org/SCONE/issues).
+
+* **Search for existing feature requests.** Please check if a similar request already exists. If so, please add any further information here.
+* **Let us know what you want to achieve.** It often helps to know the end-goal of your request. This helps us come up with better and more generic solutions, and also provide possible workarounds.
+
+## Improving the source code
+
+Before sending a pull-request, please let us know via the [SCONE GitHub issues page](https://github.com/opensim-org/SCONE/issues). We love to think along with you and get the best out of your efforts.
+
+#### Building SCONE
+
+For more information on how to build SCONE, please see the [Building Instructions](BUILD.md).
+
+#### Code Style
+
+Please try to match existing code styles when making changes. Note that code styles may differ between different library dependencies. When in doubt, ask. Eventually we might compile these questions into a separate document.
+
+#### Pull requests
+- Keep them simple and small. If possible, squash your commits.
+- Try to share how your code has been tested before submitting a pull request.
+- If your PR resolves an issue, include **closes #ISSUE_NUMBER** in your commit message (or a [synonym](https://help.github.com/articles/closing-issues-via-commit-messages)).
+
