### 2019-06-12 7fc5baa
Update CONTRIBUTING.md
