### 2019-06-12 b94baeb
Update CONTRIBUTING.md
