### 2019-06-12 3be6879
Update CONTRIBUTING.md
