### 2019-06-12 d1de354
Update CONTRIBUTING.md
