### 2019-06-12 #1421 @tgeijten
Thanks again @modenaxe for checking everything so thoroughly. I have addressed all points mentioned in your post, most of which have been added to the [SCONE FAQ](https://scone.software/doku.php?id=faq). In the [paper](https://github.com/opensim-org/SCONE/blob/master/paper/paper.md), I replaced the 'features of SCONE' list with an 'intended users' list; I believe it works better that way, let me know if you agree.

Thanks @Kevin-Mattheus-Moerman for the suggestions and examples. I have added a [CONTRIBUTING.md](https://github.com/opensim-org/SCONE/blob/master/CONTRIBUTING.md) and updated the [README.md](https://github.com/opensim-org/SCONE/blob/master/README.md), as well as some of the links on the [website](https://scone.software). I will consider adding a `CODE_OF_CONDUCT.md` later, even though I hope it won't be necessary ☺️
