### 2019-06-12 #1421 @demotu
I installed SCONE and ran some examples and tutorials. The software runs perfect and the documentation is very well written. After the additions that have been made I don't have any critics to the paper. Congratulations for all the work.
