### 2019-06-12 3e0056c
Minor change to paper to fix bullet list
