### 2019-06-12 #1421 @Kevin-Mattheus-Moerman
@tgeijten can you check the following:

- [ ] Your references, it looks like your `.bib` file can be cleaned up. Please remove references you do not use in the paper. I think I fixed it here: https://github.com/opensim-org/SCONE/pull/63
 - [ ] The first set of bullet points. I think some spaces at the end of `:` should do it. I propose it here: https://github.com/opensim-org/SCONE/pull/62
