### 2019-06-12 1b79064
Cleaning up .bib file for JOSS paper (#63)
