+++ paper/paper.bib
@@ -9,25 +9,7 @@ title = {{Dynamic Optimization of Human Walking}},
 volume = {123},
 year = {2001}
 }
-@article{Seth2018,
-author = {Seth, Ajay and Hicks, Jennifer L and Uchida, Thomas K and Habib, Ayman and Dembia, Christopher L and Dunne, James J and Ong, Carmichael F and DeMers, Matthew S. and Rajagopal, Apoorva and Millard, Matthew and Hamner, Samuel R and Arnold, Edith M and Yong, Jennifer R and Lakshmikanth, Shrinidhi K and Sherman, Michael A and Ku, Joy P and Delp, Scott L},
-doi = {10.1371/journal.pcbi.1006223},
-isbn = {1111111111},
-issn = {15537358},
-journal = {PLoS Computational Biology},
-number = {7},
-pmid = {30048444},
-title = {{OpenSim: Simulating musculoskeletal dynamics and neuromuscular control to study human and animal movement}},
-volume = {14},
-year = {2018}
-}
-@article{Hansen2006,
-author = {Hansen, Nikolaus},
-journal = {Towards a new evolutionary computation},
-pages = {75--102},
-title = {{The CMA evolution strategy: a comparing review}},
-year = {2006}
-}
+
 @article{Geijtenbeek2013,
 author = {Geijtenbeek, Thomas and van de Panne, Michiel and van der Stappen, a. Frank},
 doi = {10.1145/2508363.2508399},
