### 2019-06-12 c69f3b1
N: added get_contact_force() / get_contact_moment() to LuaBody scripting interface, closes #58
