### 2019-06-12 #1421 @Kevin-Mattheus-Moerman
@tgeijten looks like I broke both bullet lists. You need an enter between the text and the bullet lists.
