### 2019-06-12 7615ba2
N: Initial work on support for storing contact forces/moments, only works for grf though. See labels #issue55 for future work.
