+++ src/sconelib/scone/core/HasData.h
@@ -25,6 +25,7 @@ namespace scone
 		BodyComPosition,
 		BodyOrientation,
 		GroundReactionForce,
+		ContactForce,
 		ExternalForce,
 		CenterOfMass,
 		SensorData,
