### 2019-06-12 #1421 @Kevin-Mattheus-Moerman
@tgeijten at this point can you please do the following:

- [ ] Can you archive a copy of the reviewed software in [ZENODO](https://zenodo.org/)? Can you list the DOI of the archived version here when you have it? Make sure the ZENODO archive title matches your paper title and that the author list matches with the paper as well (you might have to remove some contributors e.g. people like me who made a minor contribution).
- [ ] Can you provide the version tag for the reviewed/archived package (it may have changed during review)?
