### 2019-06-13 48962fc
U: added Seth2018, Hansen2006, Geyer2010 to paper.bib
