+++ paper/paper.bib
@@ -22,3 +22,34 @@ title = {{Flexible muscle-based locomotion for bipedal creatures}},
 volume = {32},
 year = {2013}
 }
+
+@article{Seth2018,
+title={OpenSim: Simulating musculoskeletal dynamics and neuromuscular control to study human and animal movement},
+author={Seth, Ajay and Hicks, Jennifer L and Uchida, Thomas K and Habib, Ayman and Dembia, Christopher L and Dunne, James J and Ong, Carmichael F and DeMers, Matthew S and Rajagopal, Apoorva and Millard, Matthew and others},
+journal={PLoS computational biology},
+volume={14},
+number={7},
+pages={e1006223},
+year={2018},
+publisher={Public Library of Science}
+}
+
+@incollection{Hansen2006,
+title={The CMA evolution strategy: a comparing review},
+author={Hansen, Nikolaus},
+booktitle={Towards a new evolutionary computation},
+pages={75--102},
+year={2006},
+publisher={Springer}
+}
+
+@article{Geyer2010,
+title={A muscle-reflex model that encodes principles of legged mechanics produces human walking dynamics and muscle activities},
+author={Geyer, Hartmut and Herr, Hugh},
+journal={IEEE Transactions on neural systems and rehabilitation engineering},
+volume={18},
+number={3},
+pages={263--273},
+year={2010},
+publisher={IEEE}
+}
