+++ paper/paper.bib
@@ -1,26 +1,26 @@
 @article{Anderson2001,
-author = {Anderson, Frank C. and Pandy, Marcus G.},
-doi = {10.1115/1.1392310},
-issn = {01480731},
-journal = {Journal of Biomechanical Engineering},
-number = {5},
-pages = {381},
-title = {{Dynamic Optimization of Human Walking}},
-volume = {123},
-year = {2001}
+author={Anderson, Frank C. and Pandy, Marcus G.},
+doi={10.1115/1.1392310},
+issn={01480731},
+journal={Journal of Biomechanical Engineering},
+number={5},
+pages={381},
+title={{Dynamic Optimization of Human Walking}},
+volume={123},
+year={2001}
 }
 
 @article{Geijtenbeek2013,
-author = {Geijtenbeek, Thomas and van de Panne, Michiel and van der Stappen, a. Frank},
-doi = {10.1145/2508363.2508399},
-issn = {07300301},
-journal = {ACM Transactions on Graphics},
-month = {nov},
-number = {6},
-pages = {1--11},
-title = {{Flexible muscle-based locomotion for bipedal creatures}},
-volume = {32},
-year = {2013}
+author={Geijtenbeek, Thomas and van de Panne, Michiel and van der Stappen, a. Frank},
+doi={10.1145/2508363.2508399},
+issn={07300301},
+journal={ACM Transactions on Graphics},
+month={nov},
+number={6},
+pages={1--11},
+title={{Flexible muscle-based locomotion for bipedal creatures}},
+volume={32},
+year={2013}
 }
 
 @article{Seth2018,
