### 2019-06-13 8dd2988
Minor paper.md and paper.bib fixes
