+++ paper/paper.bib
@@ -1,16 +1,17 @@
 @article{Anderson2001,
+title={Dynamic Optimization of Human Walking},
 author={Anderson, Frank C. and Pandy, Marcus G.},
 doi={10.1115/1.1392310},
 issn={01480731},
 journal={Journal of Biomechanical Engineering},
 number={5},
 pages={381},
-title={{Dynamic Optimization of Human Walking}},
 volume={123},
 year={2001}
 }
 
 @article{Geijtenbeek2013,
+title={Flexible muscle-based locomotion for bipedal creatures},
 author={Geijtenbeek, Thomas and van de Panne, Michiel and van der Stappen, a. Frank},
 doi={10.1145/2508363.2508399},
 issn={07300301},
@@ -18,7 +19,6 @@ journal={ACM Transactions on Graphics},
 month={nov},
 number={6},
 pages={1--11},
-title={{Flexible muscle-based locomotion for bipedal creatures}},
 volume={32},
 year={2013}
 }
@@ -26,6 +26,7 @@ year={2013}
 @article{Seth2018,
 title={OpenSim: Simulating musculoskeletal dynamics and neuromuscular control to study human and animal movement},
 author={Seth, Ajay and Hicks, Jennifer L and Uchida, Thomas K and Habib, Ayman and Dembia, Christopher L and Dunne, James J and Ong, Carmichael F and DeMers, Matthew S and Rajagopal, Apoorva and Millard, Matthew and others},
+doi={10.1371/journal.pcbi.1006223},
 journal={PLoS computational biology},
 volume={14},
 number={7},
@@ -37,6 +38,7 @@ publisher={Public Library of Science}
 @incollection{Hansen2006,
 title={The CMA evolution strategy: a comparing review},
 author={Hansen, Nikolaus},
+doi={10.1007/3-540-32494-1_4},
 booktitle={Towards a new evolutionary computation},
 pages={75--102},
 year={2006},
@@ -46,6 +48,7 @@ publisher={Springer}
 @article{Geyer2010,
 title={A muscle-reflex model that encodes principles of legged mechanics produces human walking dynamics and muscle activities},
 author={Geyer, Hartmut and Herr, Hugh},
+doi={10.1109/TNSRE.2010.2047592},
 journal={IEEE Transactions on neural systems and rehabilitation engineering},
 volume={18},
 number={3},
