### 2019-06-13 #1421 @tgeijten
@Kevin-Mattheus-Moerman what exactly do I need to archive in ZENODO? Just the binary installer used by the reviewers, or the entire GitHub repository?
