### 2019-06-13 a674a3e
F: enum initialization of Optimizer::log_level and JumpMeasure::jump_type
