+++ src/sconelib/scone/measures/JumpMeasure.cpp
@@ -26,7 +26,7 @@ namespace scone
 		INIT_PROP( props, recover_time, 1.0 );
 		INIT_PROP( props, terminate_on_peak, true );
 		INIT_PROP( props, negate_result, false );
-		jump_type = static_cast<JumpType>( props.get<int>( "jump_type" ) );
+		jump_type = static_cast<JumpType>( props.get<int>( "jump_type", HighJump ) );
 		INIT_PROP( props, minimize, false ); // defaults to false
 		INIT_PROP( props, offset, Vec3::zero() );
 
