### 2019-06-13 7e7a868
N: Added LuaVec3 multiply, add, subtract, tostring
