### 2019-06-14 92a70a5
U: paper.md tags now match ZENODO submission, minor edits
