+++ paper/paper.md
@@ -8,7 +8,8 @@ date: "19 April 2019"
 bibliography: paper.bib
 tags:
 - predictive simulation
-- neurobiomechanical modeling
+- biomechanical modeling
+- neuromuscular control
 - optimization
 - data visualization
 affiliations:
@@ -25,11 +26,9 @@ Predictive forward dynamic simulations do the opposite: they compute motion traj
 
 Despite having shown great promise [@Anderson2001, @Geyer2010, @Geijtenbeek2013], the number of studies that successfully employed predictive forward simulations has been remarkably limited. A main contributing factor is complexity: besides expertise in neurological and musculoskeletal modeling, users require knowledge and understanding of optimization theory, as well as advanced software development skills to tie these components together. Even successful projects often do not permit fruitful collaboration and follow-up research, due to lack of structure and documentation of the resulting code-base.
 
-SCONE (https://scone.software) is designed to help overcome these obstacles. It is a fully featured software framework that allows researchers to perform, analyze and reproduce custom predictive simulations of biological motion. SCONE features a user-friendly graphical user interface and can be used without programming skills.
+SCONE (https://scone.software) is designed to help overcome these obstacles. It is a fully featured software framework that allows researchers to perform, analyze and reproduce custom predictive simulations of biological motion. SCONE features a user-friendly graphical user interface and can be used without programming skills. It is designed with a wide potential audience in mind, including:   
 
-SCONE is designed to cater to a wide range of potential users, including:   
-
-  * Clinical researchers with limited technical skills, who wish to perform *what-if* scenarios using existing SCONE Scenarios.
+  * Clinical researchers with limited technical skills, who wish to perform *what-if* scenarios using existing SCONE scenarios.
   * Biomechanics / neuromechanics researchers studying neuromuscular control.
   * Robotics researchers interested in optimized control strategies, or the interaction between humans and assistive devices.
 
@@ -48,7 +47,7 @@ Please visit https://scone.software for more information on SCONE, including tut
 # Acknowledgements
 The development of SCONE has been made possible through the following financial contributions:    
 
-  * The European Union’s Seventh Framework Programme for research, technological development and demonstration under grant agreement no 600932.
+  * The European Union's Seventh Framework Programme for research, technological development and demonstration under grant agreement no 600932.
   * The National Center for Simulation in Rehabilitation Research (NIH P2C HD065690).
   * The VENI research program with project number 15153, which is (partly) financed by the Netherlands Organisation for Scientific Research (NWO).
 
