### 2019-06-14 92c49c6
U: minor edit to paper.md
