### 2019-06-14 22d40a4
U: paper.md fix for better layout
