### 2019-06-14 #1421 @tgeijten
@Kevin-Mattheus-Moerman thanks for the information. I've uploaded the repository to ZENODO, the resulting DOI is `10.5281/zenodo.3245810`.

The latest release is version 1.1.1, but I believe the reviews are based on version 1.0.0. Both versions are available in the repository as separate branches.

Please let me know if you need anything else at this point.
