### 2019-06-14 #1421 @Kevin-Mattheus-Moerman
@tgeijten just to be sure. If the latest version (v1.1.1) contains the changes implemented during the review process please archive that version.
