### 2019-06-14 #1421 @tgeijten
@Kevin-Mattheus-Moerman sorry, I think I misread. Just to be complete:
* The archived version is the most recent (today's) version of the repository. It contains all the code in binary releases that have been tested by the reviewers (in separate branches).
* Changes suggested by reviewers are only in text (e.g. CONTRIBUTING.md) and on the website, not to the actual source code. These textual changes have only been applied to the master branch and do not have a new version number, since the version numbers coincide with binary releases (which there haven't been since these changes).

Hope that makes it a bit clearer, sorry for the confusion.
