+++ src/sconelib/scone/controllers/ExternalBalanceController.cpp
@@ -10,6 +10,7 @@
 
 #include "scone/model/Model.h"
 #include "scone/optimization/Params.h"
+#include "xo/geometry/euler_angles.h"
 
 namespace scone
 {
