### 2019-06-14 2723b6c
R: lua_api now uses xo_lua for vec3 and quat
