### 2019-06-14 #1421 @Kevin-Mattheus-Moerman
@tgeijten sorry what is still unclear is the version tag we should use here since you said

> The latest release is version 1.1.1, but I believe the reviews are based on version 1.0.0

But you also said:

> The archived version is the most recent (today's) version

However, the ZENODO archive is labelled as v1.0.0.

I get the impression review started here with v1.0.0 and that now (after review) it is at v1.1.1? Is this correct? If so we'll use v1.1.1 as the assigned version here, and also you'd need to update the metadata on ZENODO to reflect that version tag.
