### 2019-06-14 3a0c5cd
N: Added lua-5.3.5 to the repository. Reason for this is that CMake's FindLua would not detect debug libraries, which caused compiler warnings. Setting SCONE_LUA causes the library to be build into the bin folder.
