+++ contrib/lua-5.3.5/CMakeLists.txt
@@ -1,0 +1,11 @@
+find_path( LUA_INCLUDE_DIR NAMES lapi.h PATHS ${CMAKE_CURRENT_SOURCE_DIR}/src )
+
+set ( SRC_CORE src/lapi.c src/lcode.c src/lctype.c src/ldebug.c src/ldo.c src/ldump.c src/lfunc.c src/lgc.c src/llex.c
+  src/lmem.c src/lobject.c src/lopcodes.c src/lparser.c src/lstate.c src/lstring.c src/ltable.c
+  src/ltm.c src/lundump.c src/lvm.c src/lzio.c )
+set ( SRC_LIB src/lauxlib.c src/lbaselib.c src/lbitlib.c src/lcorolib.c src/ldblib.c src/liolib.c
+  src/lmathlib.c src/loslib.c src/lstrlib.c src/ltablib.c src/lutf8lib.c src/loadlib.c src/linit.c )
+
+# build static lua library
+add_library ( liblua ${SRC_CORE} ${SRC_LIB} )
+set_target_properties ( liblua PROPERTIES OUTPUT_NAME lua )
