### 2019-06-14 #1421 @tgeijten
@Kevin-Mattheus-Moerman sorry for the confusion, but version 1.1.1 also does not contain the (textual) changes suggested by reviewers. These changes are only in the master branch and have no version number assigned to them.

What do you suggest I do? The next 'official' release to incorporate all these changes (and many others) will be 1.2.0, but that's still in development. I could also create a new artificial version tag to represent the current intermediate state, but I'm not sure if that will be helpful to users.
