### 2019-06-14 #1421 @Kevin-Mattheus-Moerman
@tgeijten I got it now sorry you did say that earlier. We go with version 1.0.0 since that was reviewed, contains the requested changes, and has been archived.
